+++ backend/src/graphql/resolvers/GdtResolver.ts
@@ -5,9 +5,10 @@ import { Resolver, Query, Args, Ctx, Authorized } from 'type-graphql'
 import { getCustomRepository } from 'typeorm'
 import CONFIG from '../../config'
 import { GdtEntryList } from '../models/GdtEntryList'
-import { GdtTransactionSessionIdInput } from '../inputs/GdtInputs'
+import Paginated from '../args/Paginated'
 import { apiGet } from '../../apis/HttpRequest'
 import { UserRepository } from '../../typeorm/repository/User'
+import { Order } from '../enum/Order'
 
 @Resolver()
 export class GdtResolver {
@@ -16,7 +17,7 @@ export class GdtResolver {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   async listGDTEntries(
     @Args()
-    { currentPage = 1, pageSize = 5, order = 'DESC' }: GdtTransactionSessionIdInput,
+    { currentPage = 1, pageSize = 5, order = Order.DESC }: Paginated,
     @Ctx() context: any,
   ): Promise<GdtEntryList> {
     // load user
