+++ backend/src/graphql/resolvers/TransactionResolver.ts
@@ -3,22 +3,217 @@
 
 import { Resolver, Query, Args, Authorized, Ctx, Mutation } from 'type-graphql'
 import { getCustomRepository } from 'typeorm'
+
 import CONFIG from '../../config'
-import { TransactionList } from '../models/Transaction'
-import { TransactionListInput, TransactionSendArgs } from '../inputs/TransactionInput'
-import { apiGet, apiPost } from '../../apis/HttpRequest'
+
+import { Transaction } from '../models/Transaction'
+import { TransactionList } from '../models/TransactionList'
+
+import TransactionSendArgs from '../args/TransactionSendArgs'
+import Paginated from '../args/Paginated'
+
+import { Order } from '../enum/Order'
+
 import { BalanceRepository } from '../../typeorm/repository/Balance'
 import { UserRepository } from '../../typeorm/repository/User'
-import listTransactions from './listTransactions'
+import { UserTransactionRepository } from '../../typeorm/repository/UserTransaction'
+import { TransactionRepository } from '../../typeorm/repository/Transaction'
+
+import { User as dbUser } from '../../typeorm/entity/User'
+import { UserTransaction as dbUserTransaction } from '../../typeorm/entity/UserTransaction'
+import { Transaction as dbTransaction } from '../../typeorm/entity/Transaction'
+
+import { apiGet, apiPost } from '../../apis/HttpRequest'
 import { roundFloorFrom4 } from '../../util/round'
-import { calculateDecay } from '../../util/decay'
+import { calculateDecay, calculateDecayWithInterval } from '../../util/decay'
+import { TransactionTypeId } from '../enum/TransactionTypeId'
+import { TransactionType } from '../enum/TransactionType'
+
+// Helper function
+async function calculateAndAddDecayTransactions(
+  userTransactions: dbUserTransaction[],
+  user: dbUser,
+  decay: boolean,
+  skipFirstTransaction: boolean,
+): Promise<Transaction[]> {
+  const finalTransactions: Transaction[] = []
+  const transactionIds: number[] = []
+  const involvedUserIds: number[] = []
+
+  userTransactions.forEach((userTransaction: dbUserTransaction) => {
+    transactionIds.push(userTransaction.transactionId)
+  })
+
+  const transactionRepository = getCustomRepository(TransactionRepository)
+  const transactions = await transactionRepository.joinFullTransactionsByIds(transactionIds)
+
+  const transactionIndiced: dbTransaction[] = []
+  transactions.forEach((transaction: dbTransaction) => {
+    transactionIndiced[transaction.id] = transaction
+    if (transaction.transactionTypeId === 2) {
+      involvedUserIds.push(transaction.transactionSendCoin.userId)
+      involvedUserIds.push(transaction.transactionSendCoin.recipiantUserId)
+    }
+  })
+  // remove duplicates
+  // https://stackoverflow.com/questions/1960473/get-all-unique-values-in-a-javascript-array-remove-duplicates
+  const involvedUsersUnique = involvedUserIds.filter((v, i, a) => a.indexOf(v) === i)
+  const userRepository = getCustomRepository(UserRepository)
+  const userIndiced = await userRepository.getUsersIndiced(involvedUsersUnique)
+
+  const decayStartTransaction = await transactionRepository.findDecayStartBlock()
+
+  for (let i = 0; i < userTransactions.length; i++) {
+    const userTransaction = userTransactions[i]
+    const transaction = transactionIndiced[userTransaction.transactionId]
+    const finalTransaction = new Transaction()
+    finalTransaction.transactionId = transaction.id
+    finalTransaction.date = transaction.received.toISOString()
+    finalTransaction.memo = transaction.memo
+    finalTransaction.totalBalance = roundFloorFrom4(userTransaction.balance)
+    const prev = i > 0 ? userTransactions[i - 1] : null
+
+    if (prev && prev.balance > 0) {
+      const current = userTransaction
+      const decay = await calculateDecayWithInterval(
+        prev.balance,
+        prev.balanceDate,
+        current.balanceDate,
+      )
+      const balance = prev.balance - decay.balance
+
+      if (balance) {
+        finalTransaction.decay = decay
+        finalTransaction.decay.balance = roundFloorFrom4(balance)
+        if (
+          decayStartTransaction &&
+          prev.transactionId < decayStartTransaction.id &&
+          current.transactionId > decayStartTransaction.id
+        ) {
+          finalTransaction.decay.decayStartBlock = (
+            decayStartTransaction.received.getTime() / 1000
+          ).toString()
+        }
+      }
+    }
+
+    // sender or receiver when user has sended money
+    // group name if creation
+    // type: gesendet / empfangen / geschöpft
+    // transaktion nr / id
+    // date
+    // balance
+    if (userTransaction.transactionTypeId === TransactionTypeId.CREATION) {
+      // creation
+      const creation = transaction.transactionCreation
+
+      finalTransaction.name = 'Gradido Akademie'
+      finalTransaction.type = TransactionType.CREATION
+      // finalTransaction.targetDate = creation.targetDate
+      finalTransaction.balance = roundFloorFrom4(creation.amount)
+    } else if (userTransaction.transactionTypeId === TransactionTypeId.SEND) {
+      // send coin
+      const sendCoin = transaction.transactionSendCoin
+      let otherUser: dbUser | undefined
+      finalTransaction.balance = roundFloorFrom4(sendCoin.amount)
+      if (sendCoin.userId === user.id) {
+        finalTransaction.type = TransactionType.SEND
+        otherUser = userIndiced[sendCoin.recipiantUserId]
+        // finalTransaction.pubkey = sendCoin.recipiantPublic
+      } else if (sendCoin.recipiantUserId === user.id) {
+        finalTransaction.type = TransactionType.RECIEVE
+        otherUser = userIndiced[sendCoin.userId]
+        // finalTransaction.pubkey = sendCoin.senderPublic
+      } else {
+        throw new Error('invalid transaction')
+      }
+      if (otherUser) {
+        finalTransaction.name = otherUser.firstName + ' ' + otherUser.lastName
+        finalTransaction.email = otherUser.email
+      }
+    }
+    if (i > 0 || !skipFirstTransaction) {
+      finalTransactions.push(finalTransaction)
+    }
+
+    if (i === userTransactions.length - 1 && decay) {
+      const now = new Date()
+      const decay = await calculateDecayWithInterval(
+        userTransaction.balance,
+        userTransaction.balanceDate,
+        now.getTime(),
+      )
+      const balance = userTransaction.balance - decay.balance
+      if (balance) {
+        const decayTransaction = new Transaction()
+        decayTransaction.type = 'decay'
+        decayTransaction.balance = roundFloorFrom4(balance)
+        decayTransaction.decayDuration = decay.decayDuration
+        decayTransaction.decayStart = decay.decayStart
+        decayTransaction.decayEnd = decay.decayEnd
+        finalTransactions.push(decayTransaction)
+      }
+    }
+  }
+
+  return finalTransactions
+}
+
+// Helper function
+async function listTransactions(
+  currentPage: number,
+  pageSize: number,
+  order: Order,
+  user: dbUser,
+): Promise<TransactionList> {
+  let limit = pageSize
+  let offset = 0
+  let skipFirstTransaction = false
+  if (currentPage > 1) {
+    offset = (currentPage - 1) * pageSize - 1
+    limit++
+  }
+
+  if (offset && order === Order.ASC) {
+    offset--
+  }
+  const userTransactionRepository = getCustomRepository(UserTransactionRepository)
+  let [userTransactions, userTransactionsCount] = await userTransactionRepository.findByUserPaged(
+    user.id,
+    limit,
+    offset,
+    order,
+  )
+  skipFirstTransaction = userTransactionsCount > offset + limit
+  const decay = !(currentPage > 1)
+  let transactions: Transaction[] = []
+  if (userTransactions.length) {
+    if (order === Order.DESC) {
+      userTransactions = userTransactions.reverse()
+    }
+    transactions = await calculateAndAddDecayTransactions(
+      userTransactions,
+      user,
+      decay,
+      skipFirstTransaction,
+    )
+    if (order === Order.DESC) {
+      transactions = transactions.reverse()
+    }
+  }
+
+  const transactionList = new TransactionList()
+  transactionList.count = userTransactionsCount
+  transactionList.transactions = transactions
+  return transactionList
+}
 
 @Resolver()
 export class TransactionResolver {
   @Authorized()
   @Query(() => TransactionList)
   async transactionList(
-    @Args() { firstPage = 1, items = 25, order = 'DESC' }: TransactionListInput,
+    @Args() { currentPage = 1, pageSize = 25, order = Order.DESC }: Paginated,
     @Ctx() context: any,
   ): Promise<TransactionList> {
     // get public key for current logged in user
@@ -29,7 +224,7 @@ export class TransactionResolver {
     const userRepository = getCustomRepository(UserRepository)
     const userEntity = await userRepository.findByPubkeyHex(result.data.user.public_hex)
 
-    const transactions = await listTransactions(firstPage, items, order, userEntity)
+    const transactions = await listTransactions(currentPage, pageSize, order, userEntity)
 
     // get gdt sum
     const resultGDTSum = await apiPost(`${CONFIG.GDT_API_URL}/GdtEntries/sumPerEmailApi`, {
