+++ backend/src/graphql/resolvers/UserResolver.ts
@@ -9,13 +9,11 @@ import { SendPasswordResetEmailResponse } from '../models/SendPasswordResetEmail
 import { UpdateUserInfosResponse } from '../models/UpdateUserInfosResponse'
 import { User } from '../models/User'
 import encode from '../../jwt/encode'
-import {
-  ChangePasswordArgs,
-  CheckUsernameArgs,
-  CreateUserArgs,
-  UnsecureLoginArgs,
-  UpdateUserInfosArgs,
-} from '../inputs/LoginUserInput'
+import ChangePasswordArgs from '../args/ChangePasswordArgs'
+import CheckUsernameArgs from '../args/CheckUsernameArgs'
+import CreateUserArgs from '../args/CreateUserArgs'
+import UnsecureLoginArgs from '../args/UnsecureLoginArgs'
+import UpdateUserInfosArgs from '../args/UpdateUserInfosArgs'
 import { apiPost, apiGet } from '../../apis/HttpRequest'
 import {
   klicktippRegistrationMiddleware,
